requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
jinja2==3.1.4
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.49.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
apscheduler==3.10.4
pyyaml==6.0.2
pypdf==5.1.0
python-dotenv==1.0.1
